services/content-intelligence/pyproject.toml
[project]
name = "content-intelligence-service"
version = "1.0.0"
description = "Multi-stage retrieval with hybrid search and cross-encoder reranking"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pydantic>=2.10.0",
    "sentence-transformers>=3.0.0",
    "faiss-cpu>=1.9.0",
    "rank-bm25>=0.2.2",
    "networkx>=3.4",
    "tree-sitter>=0.23.0",
    "tree-sitter-python>=0.23.0",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cis"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
